+++ CHANGES.md
@@ -36,6 +36,7 @@ Changes in 12.2
 - Use FinalGradesQtrOrProCalculate() function in GradebookBreakdown.php
 - Fix FinalGradesGetAssignmentsPoints() when $mp_id is not UserMP() in FinalGrades.inc.php
 - CSS themes small fixes + Prevent body from scrolling when help open in stylesheet.css
+- SQL avoid query, get GRADE_LETTER from FinalGradesQtrOrProCalculate() in Grades.php, GradebookBreakdown.php & FinalGrades.inc.php
 
 Changes in 12.1.3
 -----------------